+++ src/main/java/org/pollub/campusmate/service/UserService.java
@@ -2,6 +2,7 @@ package org.pollub.campusmate.service;
 
 import lombok.AllArgsConstructor;
 import org.pollub.campusmate.entity.User;
+import org.pollub.campusmate.exception.UserNotFound;
 import org.pollub.campusmate.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
@@ -14,24 +15,26 @@ public class UserService {
 
     private final UserRepository userRepository;
 
-    //TODO: exception handling
-
-    public User getUser(Long userId){
-        return userRepository.findById(userId).orElse(null);
+    public User getUser(Long userId) {
+        return userRepository.findById(userId)
+                .orElseThrow(() -> new UserNotFound("User with id " + userId + " not found"));
     }
+
 
     public User addUser(User user){
         return userRepository.save(user);
     }
 
     public void deleteUser(Long userId){
-//        if(!userRepository.existsById(userId)){
-//
-//        }
+        if(!userRepository.existsById(userId)){
+            throw new UserNotFound("Cannot execute delete operation. User with id " + userId + " not found");
+        }
         userRepository.deleteById(userId);
     }
 
     //TODO: update user method @Transactional
+
+    //TODO: List<User> foundUsers = (List<User>) userRepository.findAll();
 
     public List<User> getAllUsers(){
         Iterable<User> users = userRepository.findAll();
@@ -40,7 +43,7 @@ public class UserService {
         users.forEach(foundUsers::add);
 
         if(foundUsers.isEmpty()){
-            return null;
+            throw new UserNotFound("No users found");
         }
 
         return foundUsers;
